From: Hasso Tepper <hasso@estpak.ee> Subject: [zebra 17365] [PATCH] Ospf area id's DECIMAL -> ADDRESS
It was annoying for me to view area id's like 335544330 (20.0.0.10 in
address format) in configuration. Most of other network devices are
using address-like id's and so does zebra when using "show ip ospf *"
commands.

You can still enter id's in decimal format, but they are just shown in
address format.