From: Andrew J. Schorr <aschorr@telemetry-investments.com> Subject: [zebra 12406] patch for ripd to support "passive-interface default" behavior
In Cisco IOS, the router rip passive-interface subcommand can accept
an interface name of "default".  In that cases, all interfaces are set
to be passive by default, and you must explicitly make it non-passive
by using the "no passive-interface <ifname>" command if you want updates
to go to that interface.

<command implemented for zebra>