rework
clarify make (BSD make not working is a bug, and GNU make not working
is a really serious bug)

declare POSIX.2 as a semi-requirement, with notion that workarounds
for non-posix must be clean.

list OS versions for which not working is a bug.

add instructions for using info and emacs/info, and remove admonition
to use pinfo

add note about GNU awk (really needed?  BSD awk)

Note that NetBSD 2.99.15 has texinfo 4.7