[ospfd/BSD] add missing UINT16_MAX - picked up from FreeBSD ports patch
Should probably have AC_TYPE_UINT16_T in configure.ac to make it really
clean...  and a fresh import of gnulib, the existing copy of gnulib seems to
be manually put some stuff in zebra.h.  I'm leaving that alone for now,
since a fresh import will probably clean up a lot of cruft and it would be a
duplicated effort to fix some things now.

The FreeBSD port 0.99.11_1 files/patch-ospfd-ospfd.h can probably go away
after this is applied.