[bgpd] Avoid zombie accepted peer entries
Currently, when accepting the connection, it can be left as zombie, when the
peer just initiates a connection, but never sends data (and the TCP
connection end packets are lost).  This happens because for accepted
connections a temporary new peer entry is created until OPEN message is
exchanged, and this temporary peer entry does not get the hold time
parameter set at all.

* bgp_network.c: (bgp_accept) Set hold time and keepalive values for ACCEPT
  peers.