ospfd: make local functions static
This code is only used one place and can be made local.
Gcc is smart enough to inline local functions if it wants to.

The function also  has a big chunk of compatiablity code that
is no longer used; since quagga is now in a version control system
the source does not need to be used as a historical reference.