isisd: send proper LSP after DIS election
After an IS has been elected as the Designated IS for a LAN, it did
not refresh the content of the pseudo-node after a new node has been
connected to the same LAN. Instead, the periodically reoriginated
pseudo-node LSP still contained only those IS neighbors that were
already present when the DIS election process was commenced.

The fix for the problem schedules an LSP regeneration rather than
just reoriginating the same LSP with the old content.

Signed-off-by: Fritz Reichmann <fritz@reichmann.nl>